requirements.txt
black==22.6.0
certifi==2022.6.15
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.0
click==8.1.3
cloudinary==1.29.0
colorama==0.4.5
cycler==0.11.0
dnspython==2.2.1
Flask==2.1.2
Flask-Cors==3.0.10
fonttools==4.34.4
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
jsonpickle==2.2.0
kiwisolver==1.4.4
llvmlite==0.38.1
MarkupSafe==2.1.1
matplotlib==3.5.2
more-itertools==8.13.0
music21==7.3.3
mypy-extensions==0.4.3
numba==0.55.2
numpy==1.22.4
packaging==21.3
pathspec==0.9.0
Pillow==9.2.0
platformdirs==2.5.2
pycparser==2.21
pyparsing==3.0.9
PySoundFile==0.9.0.post1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.1
six==1.16.0
tomli==2.0.1
urllib3==1.26.9
webcolors==1.12
Werkzeug==2.1.2
WTForms==3.0.1
zipp==3.8.0
